backend/requirements.txt
# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
asyncpg==0.30.0
alembic==1.16.2

# HTTP Client
httpx==0.28.1

# PDF Processing
pypdf==5.5.0
reportlab==4.4.1

# Scheduling
apscheduler==3.11.0

# Utilities
python-dotenv==1.1.0
pydantic==2.11.5
pydantic-settings==2.9.1
